Move `Validation` protocol and impls to `validate` module
