+++ kernels/src/kernels/layer/func.py
@@ -10,7 +10,6 @@ from kernels_data import KernelDependency, KernelLocks
 from kernels.resolver import LockedHubCacheResolver, LockedHubResolver
 
 from .._versions import select_revision_or_version
-from ..deps import AllValidator, default_validators
 from ..hf_hub import _get_hf_api
 from ..load import (
     get_kernel,
@@ -21,6 +20,7 @@ from ..locking import (
     get_caller_locked_kernel_revision,
     get_locked_kernel_revision,
 )
+from ..validate import AllValidator, default_validators
 from .layer import _create_func_module, use_kernel_forward_from_hub
 from .repos import RepositoryProtocol
