+++ nix-builder/pkgs/get-kernel-check/get-kernel-check-hook.py
@@ -1,10 +1,10 @@
 import os
 from pathlib import Path
 
-from kernels.deps import AllValidator, default_validators
 from kernels.hf_hub import _get_hf_api
 from kernels.load import get_kernel_with_resolver
 from kernels.resolver import KernelPathsResolver, RepoPathsResolver, SequentialResolver
+from kernels.validate import AllValidator, default_validators
 from kernels_data import KernelDependency, KernelPaths, KernelVersion
 
 out = os.getenv("out")
